+++ src/cc/Hypertable/RangeServer/CellCache.h
@@ -62,7 +62,7 @@ namespace Hypertable {
 
     virtual void get_rows(std::vector<std::string> &rows);
 
-    virtual uint32_t get_total_entries() { return m_cell_map.size(); }
+    virtual int64_t get_total_entries() { return m_cell_map.size(); }
 
     /** Creates a CellCacheScanner object that contains an shared pointer
      * (intrusive_ptr) to this CellCache.
